(refactor) moved threbox styles override to a separate file
